(#2) 스팟찍기 레이어 UI변경에 따른 컴포넌트 구조 변경
