+++ components/Home/SpotGenerator/AreaEmoji/AreaEmoji.tsx
@@ -1,0 +1,37 @@
+import React from 'react';
+import emojis from '../../../../constants/emojis';
+import * as $ from './AreaEmojiView';
+import type { Emoji } from '../../../../constants/emojis';
+
+const AreaEmoji: React.FC = () => {
+  // @see Flex로 Flexible Grid 정렬을 하기 위한, 더미 아이템 요소를 4n개에 맞게 추가
+  const emojisToGrid: Emoji[] = [null, null, null, null].reduce(
+    (emojisToGrid, dummy) => {
+      if (emojisToGrid.length % 4 !== 0) emojisToGrid.push(dummy);
+      return emojisToGrid;
+    },
+    emojis
+  );
+  const handleClickEmoji = (e) => {
+    e.preventDefault();
+    // TODO: Emoji 선택
+  };
+
+  return (
+    <$.AreaEmoji>
+      <$.EmojiList>
+        {emojisToGrid.map((emoji, i) => (
+          <$.EmojiItem key={`emoji-button-${i}`}>
+            {emoji && (
+              <$.EmojiButton onClick={handleClickEmoji}>
+                <$.EmojiImage src={emoji.imageUrl} />
+              </$.EmojiButton>
+            )}
+          </$.EmojiItem>
+        ))}
+      </$.EmojiList>
+    </$.AreaEmoji>
+  );
+};
+
+export default AreaEmoji;
