--- components/Home/SpotGenerator/PageForm/index.ts
@@ -1,1 +1,0 @@
-export { default } from './PageForm';
