--- components/Home/SpotGenerator/PageSticker/index.ts
@@ -1,1 +1,0 @@
-export { default } from './PageSticker';
