+++ components/Home/SpotGenerator/SpotGenerator.tsx
@@ -1,8 +1,9 @@
-import { useRecoilValue, useSetRecoilState } from 'recoil';
+import { useSetRecoilState } from 'recoil';
+import AreaEmoji from './AreaEmoji';
+import AreaFilter from './AreaFilter';
+import AreaForm from './AreaForm';
 import Title from './Title';
-import PageSticker from './PageSticker';
-import PageForm from './PageForm';
-import { currentPlaceIdState, pageState } from './SpotGeneratorState';
+import { currentPlaceIdState } from './SpotGeneratorState';
 import * as $ from './SpotGeneratorView';
 import { useEffect } from 'react';
 
@@ -11,7 +12,6 @@ type Props = {
 };
 
 const SpotGenerator: React.FC<Props> = ({ placeId }) => {
-  const page = useRecoilValue(pageState);
   const setCurrentPlaceId = useSetRecoilState(currentPlaceIdState);
 
   useEffect(() => {
@@ -21,7 +21,17 @@ const SpotGenerator: React.FC<Props> = ({ placeId }) => {
   return (
     <$.SpotGenerator>
       <Title />
-      {page === 2 ? <PageForm /> : <PageSticker />}
+      <$.Inner>
+        <$.AreaText>
+          <$.HelpTitle>스티커를 원하는 장소에 붙여보세요</$.HelpTitle>
+          <$.HelpText>
+            무드 탭을 선택해 스티커로 장소의 무드를 표현해보세요
+          </$.HelpText>
+        </$.AreaText>
+        <AreaFilter />
+        <AreaEmoji />
+        <AreaForm />
+      </$.Inner>
     </$.SpotGenerator>
   );
 };
